코드 정리: idea 스크립트 problems 반영

+++ snippets.sh
@@ -1,18 +1,18 @@
 
 
 for file in *; do
-  md5=$(md5sum "$file" | awk '{print $1}')
+  md5=$(md5sum "${file}" | awk '{print $1}')
   ext="${file##*.}"
   newFile="$md5.$ext"
-  mv "$file" $newFile
+  mv "${file}" "${newFile}"
 done
 
 list=(a.ext)
-for file in ${list[*]}; do
-  md5=$(md5sum "$file" | awk '{print $1}')
+for file in "${list[@]}"; do
+  md5=$(md5sum "${file}" | awk '{print $1}')
   ext="${file##*.}"
   newFile="$md5.$ext"
-  mv "$file" $newFile
+  mv "${file}" "${newFile}"
 done
 
 
@@ -20,10 +20,10 @@ ext="ext"
 for file in *; do
   fileNm="${file%.*}"
   newFile="$fileNm.$ext"
-  mkvmerge -o $newFile "$file"
+  mkvmerge -o "${newFile}" "${file}"
 done
 
-for file in $(ls *[SHANA]*); do
-  fileNm="${file/\[SHANA\]/}"
-  rm -rf $fileNm
+for file in *'[SHANA]'*; do
+  fileNm="${file//\[SHANA\]/}"
+  rm -rf "${fileNm:?}"
 done
